Fix default project reload
