+++ src/models/projectCreator.js
@@ -1,7 +1,3 @@
-// eslint-disable-next-line import/no-cycle
-import { storage } from './storage';
-import { dom } from './DOM';
-
 const createProject = function createProject(name, pocket = []) {
   return {
     name,
@@ -9,15 +5,6 @@ const createProject = function createProject(name, pocket = []) {
   };
 };
 
-const createDefaultProject = function createDefaultProject() {
-  const newProject = createProject('Default Project');
-  const ourStore = storage();
-  ourStore.addProject(newProject);
-  document.getElementById('projectForm').reset();
-  dom().showProject();
-};
-
 export {
-  createProject,
-  createDefaultProject,
+  createProject as default,
 };
